+++ dotnet/GraphQL/Query.cs
@@ -1,20 +1,24 @@
 using GraphQL;
 using GraphQL.Types;
+using PackingOptimization.GraphQL.Types;
+using PackingOptimization.Data;
 using System;
+
 namespace PackingOptimization.GraphQL
 {
     [GraphQLMetadata("Query")]
     public class Query : ObjectGraphType<object>
     {
-        public Query()
+        public Query(IMerchantSettingsRepository _merchantSettingsRepository)
         {
             Name = "Query";
 
-            Field<StringGraphType>(
-                "test",
-                resolve: context =>
+            FieldAsync<AppSettingsType>(
+                "getAppSettings",
+                resolve: async context =>
                 {
-                    return "result";
+                    return await context.TryAsyncResolve(
+                        async c => await _merchantSettingsRepository.GetMerchantSettings());
                 }
             );
         }