+++ dotnet/data/IMerchantSettingsRepository.cs
@@ -1,0 +1,12 @@
+﻿namespace PackingOptimization.Data
+{
+    using PackingOptimization.Models;
+    using System.Threading.Tasks;
+
+    public interface IMerchantSettingsRepository
+    {
+        Task<MerchantSettings> GetMerchantSettings();
+
+        Task<bool> SetMerchantSettings(MerchantSettings merchantSettings);
+    }
+}