+++ react/typings/types.d.ts
@@ -1,0 +1,7 @@
+interface Container {
+  id: number
+  length: number
+  width: number
+  height: number
+  description: string
+}