+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added admin panel for box creation
+- Enabled Packing API to use created boxes
+
 ## [0.2.0] - 2022-04-28
 
 ### Added